@@ -8,7 +8,7 @@ import 'package:http/http.dart' as http;
 /// Implements the /api/users.identity call
 Future<UserIdentity> getUserIdentity(String accessToken) async {
   String endpoint = "https://slack.com/api/users.identity?token=$accessToken";
-  final http.Response response = await http.get(endpoint);
+  final http.Response response = await http.get(Uri.parse(endpoint));
 
   return new UserIdentity.fromMap(json.decode(response.body));
 }
@@ -16,7 +16,7 @@ Future<UserIdentity> getUserIdentity(String accessToken) async {
 /// Implements the /api/users.list call
 Future<UserList> getUsers(String accessToken) async {
   String endpoint = "https://slack.com/api/users.list?token=$accessToken";
-  final http.Response response = await http.get(endpoint);
+  final http.Response response = await http.get(Uri.parse(endpoint));
 
   return new UserList.fromMap(json.decode(response.body));
 }
